requirements.txt
alpaca-trade-api==1.2.1
apprise==0.9.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
dataclasses-json==0.5.3
dateparser==1.0.0
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
greenlet==1.1.0
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.20.0
oauthlib==3.1.0
pandas==1.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
python-binance==0.7.10
python-dateutil==2.8.1
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.0.4
python-twitter-v2==0.3.3
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.0.0
service-identity==21.1.0
six==1.15.0
SQLAlchemy==1.3.23
sqlitedict==1.7.0
stringcase==1.2.0
svgwrite==1.4
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
typing-inspect==0.6.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websocket-client==0.59.0
websockets==8.1
Werkzeug==2.0.1
zope.interface==5.4.0
